Added new ways of connecting images
